Create the device details page base layout
