+++ src/js/views/deviceDetails/style.js
@@ -1,0 +1,28 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+export default makeStyles(theme => ({
+  container: {
+    display: 'flex',
+    flexDirection: 'column',
+    height: '100%',
+  },
+  content: {
+    flex: 1,
+  },
+  dataGroup: {
+    background: '#f2f2f2',
+    border: `1px solid ${theme.palette.divider}`,
+    borderBottom: 'none',
+    marginBottom: '1rem',
+  },
+  dataGroupTitleIcon: {
+    minWidth: 'auto',
+    marginRight: theme.spacing(2),
+  },
+  tableCellBold: {
+    fontWeight: 'bold',
+  },
+  tableCellSecondary: {
+    color: theme.palette.text.secondary,
+  },
+}));
