+++ src/js/views/deviceDetails/translations/en.deviceDetails.i18n.json
@@ -1,0 +1,8 @@
+{
+  "title": "Device Details",
+  "lastUpdate": {
+    "date": "Date",
+    "key": "Key",
+    "value": "Value"
+  }
+}
