+++ src/js/views/deviceDetails/translations/pt_br.deviceDetails.i18n.json
@@ -1,0 +1,8 @@
+{
+  "title": "Detalhamento do Dispositivo",
+  "lastUpdate": {
+    "date": "Data",
+    "key": "Chave",
+    "value": "Valor"
+  }
+}
