+++ src/js/views/devices/translations/en.devices.i18n.json
@@ -8,7 +8,6 @@
   "removeFromFavoriteTooltip": "Remove from Favorites",
   "labelDisplayedRows": "{{from}}-{{to}} of {{count}}",
   "labelRowsPerPage": "Items per page:",
-  "detailsModalTitle": "Device Details ({{label}})",
   "massActions": "Mass Actions:",
   "createCertificates": "Create Certificates",
   "favorite": "Add to Favorites",
@@ -16,11 +15,6 @@
   "numberOfSelectedDevices_plural": "{{count}} Lines Selected",
   "emptyListMessage": "No Device Found",
   "createNewDevice": "Create New Device",
-  "lastUpdate": {
-    "date": "Date",
-    "key": "Key",
-    "value": "Value"
-  },
   "dataTableHead": {
     "id": "ID",
     "label": "Name",
